Move `ControlPointList` to own file
